chore(test): extract websocket probe fixture and track raw sockets
Move the child probe into a typechecked module that reuses tsconfig.vite.json
for `~/`, and destroy pipelined raw sockets in afterEach if waitUntil rejects.
