+++ frontend/server/websocket.server.errors.test.ts
@@ -3,7 +3,6 @@ import { existsSync } from "node:fs";
 import { mkdtemp, rm } from "node:fs/promises";
 import http from "node:http";
 import net from "node:net";
-import type { AddressInfo } from "node:net";
 import { tmpdir } from "node:os";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
@@ -18,6 +17,17 @@ import {
   reportBrowserSocketError,
 } from "./websocket-policy";
 import { initializeWebsocketServer, type WebsocketServerDependencies } from "./websocket.server";
+import {
+  closeServer,
+  connectClient,
+  deferred,
+  listenOnLoopback,
+  oversizedMaskedBinaryFrame,
+  waitForClose,
+  waitForOpen,
+  waitUntil,
+  websocketUpgradeRequest,
+} from "./websocket-test-helpers";
 
 vi.hoisted(() => {
   process.env["SESSION_KEY"] ??= "1234-in-process-session-key";
@@ -27,270 +37,9 @@ vi.hoisted(() => {
 
 const FRONTEND_ROOT = fileURLToPath(new URL("..", import.meta.url));
 const REPO_ROOT = fileURLToPath(new URL("../..", import.meta.url));
+const PROBE_PATH = fileURLToPath(new URL("./__fixtures__/websocket-probe.ts", import.meta.url));
 const DUMMY_API_KEY = "1234-dummy-frontend-backend-api-key";
 const DUMMY_SESSION_KEY = "1234-dummy-session-key";
-
-const PROBE_SOURCE = `
-import { register } from "node:module";
-import http from "node:http";
-import net from "node:net";
-import path from "node:path";
-import { pathToFileURL } from "node:url";
-import { WebSocket, WebSocketServer } from "ws";
-
-const frontendRoot = process.env.FRONTEND_ROOT;
-const mode = process.env.WEBSOCKET_PROBE_MODE;
-const appHref = pathToFileURL(path.join(frontendRoot, "app")).href + "/";
-register("data:text/javascript," + encodeURIComponent(\`
-export async function resolve(specifier, context, nextResolve) {
-  if (specifier.startsWith("~/")) {
-    return nextResolve(new URL(specifier.slice(2), \${JSON.stringify(appHref)}).href, context);
-  }
-  return nextResolve(specifier, context);
-}
-\`));
-
-const [{ initializeWebsocketServer }, policy, { attachWebsocketServerErrorListener }] = await Promise.all([
-  import(pathToFileURL(path.join(frontendRoot, "server/websocket.server.ts")).href),
-  import(pathToFileURL(path.join(frontendRoot, "server/websocket-policy.ts")).href),
-  import(pathToFileURL(path.join(frontendRoot, "server/http-server-lifecycle.ts")).href),
-]);
-const { MAX_WEBSOCKET_PAYLOAD_BYTES, attachBrowserWebsocketErrorListener, reportBrowserSocketError, errorCode } = policy;
-
-function deferred() {
-  let resolve;
-  const promise = new Promise((res) => { resolve = res; });
-  return { promise, resolve };
-}
-
-function listenOnLoopback(server) {
-  return new Promise((resolve, reject) => {
-    server.once("error", reject);
-    server.listen(0, "127.0.0.1", () => resolve(server.address()));
-  });
-}
-
-function waitForOpen(ws, timeoutMs = 5000) {
-  if (ws.readyState === WebSocket.OPEN) return Promise.resolve();
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => reject(new Error("open timeout")), timeoutMs);
-    ws.once("open", () => { clearTimeout(timer); resolve(); });
-  });
-}
-
-function waitForClose(ws, timeoutMs = 5000) {
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => reject(new Error("close timeout")), timeoutMs);
-    ws.once("close", (code) => { clearTimeout(timer); resolve(code); });
-  });
-}
-
-function websocketUpgradeRequest(port) {
-  const key = Buffer.from("1234567890abcdef").toString("base64");
-  return Buffer.from(
-    "GET /ws HTTP/1.1\\r\\n" +
-    "Host: 127.0.0.1:" + port + "\\r\\n" +
-    "Upgrade: websocket\\r\\n" +
-    "Connection: Upgrade\\r\\n" +
-    "Sec-WebSocket-Key: " + key + "\\r\\n" +
-    "Sec-WebSocket-Version: 13\\r\\n" +
-    "\\r\\n",
-  );
-}
-
-function oversizedMaskedBinaryFrame(payloadLength) {
-  const header = Buffer.alloc(2 + 8 + 4);
-  header[0] = 0x82;
-  header[1] = 0x80 | 127;
-  header.writeBigUInt64BE(BigInt(payloadLength), 2);
-  return Buffer.concat([header, Buffer.alloc(payloadLength)]);
-}
-
-const auth = deferred();
-let acceptedErrorCode = null;
-let resolveAccepted = () => {};
-const accepted = new Promise((resolve) => { resolveAccepted = resolve; });
-
-function report(error, context) {
-  reportBrowserSocketError(error, context);
-  const code = errorCode(error);
-  process.stdout.write("ACCEPTED_ERROR_CODE=" + code + "\\n");
-  if (acceptedErrorCode == null) {
-    acceptedErrorCode = code;
-    resolveAccepted(code);
-  }
-}
-
-const httpServer = http.createServer();
-const wss = new WebSocketServer({
-  server: httpServer,
-  path: "/ws",
-  maxPayload: MAX_WEBSOCKET_PAYLOAD_BYTES,
-});
-attachWebsocketServerErrorListener(wss, {
-  isOwned: () => false,
-  onUnexpectedError: (error) => {
-    process.stderr.write("WSS_FATAL " + error.message + "\\n");
-    process.exitCode = 1;
-    process.exit(1);
-  },
-});
-initializeWebsocketServer(wss, {
-  authenticate: () => auth.promise,
-  startBackendClient: () => {},
-  reportBrowserSocketError: report,
-  registerBrowserSocketErrorListener: mode === "oversized-pre-auth-listener-removed"
-    ? () => {}
-    : attachBrowserWebsocketErrorListener,
-});
-
-const address = await listenOnLoopback(httpServer);
-const url = "ws://127.0.0.1:" + address.port + "/ws";
-let clientCloseCode = null;
-let nextConnectionOpened = false;
-
-try {
-  if (mode === "pipelined-oversized") {
-    const socket = net.connect(address.port, "127.0.0.1");
-    socket.on("error", () => {});
-    await new Promise((resolve, reject) => {
-      socket.once("connect", resolve);
-      socket.once("error", reject);
-    });
-    socket.write(Buffer.concat([
-      websocketUpgradeRequest(address.port),
-      oversizedMaskedBinaryFrame(MAX_WEBSOCKET_PAYLOAD_BYTES + 1),
-    ]));
-    await Promise.race([
-      accepted,
-      new Promise((_, reject) => setTimeout(() => reject(new Error("accepted error timeout")), 8000)),
-    ]);
-    socket.destroy();
-  } else {
-    const client = new WebSocket(url);
-    client.on("error", () => {});
-    await waitForOpen(client);
-    const closed = waitForClose(client);
-    client.send(Buffer.alloc(MAX_WEBSOCKET_PAYLOAD_BYTES + 1));
-    clientCloseCode = await closed;
-    await Promise.race([
-      accepted,
-      new Promise((_, reject) => setTimeout(() => reject(new Error("accepted error timeout")), 8000)),
-    ]);
-  }
-
-  const healthy = new WebSocket(url);
-  healthy.on("error", () => {});
-  await waitForOpen(healthy);
-  nextConnectionOpened = true;
-  healthy.close();
-  await waitForClose(healthy);
-
-  const observation = { acceptedErrorCode, clientCloseCode, nextConnectionOpened };
-  process.stdout.write("PROBE_RESULT=" + JSON.stringify(observation) + "\\n");
-  process.stdout.write("probe-complete\\n");
-  process.exitCode = 0;
-} finally {
-  auth.resolve(false);
-  for (const client of wss.clients) client.terminate();
-  await new Promise((resolve) => wss.close(() => resolve()));
-  await new Promise((resolve) => httpServer.close(() => resolve()));
-}
-`;
-
-function deferred<T>(): {
-  promise: Promise<T>;
-  resolve: (value: T) => void;
-  reject: (reason?: unknown) => void;
-} {
-  let resolve!: (value: T) => void;
-  let reject!: (reason?: unknown) => void;
-  const promise = new Promise<T>((res, rej) => {
-    resolve = res;
-    reject = rej;
-  });
-  return { promise, resolve, reject };
-}
-
-function listenOnLoopback(server: http.Server): Promise<AddressInfo> {
-  return new Promise((resolve, reject) => {
-    server.once("error", reject);
-    server.listen(0, "127.0.0.1", () => {
-      const address = server.address();
-      if (!address || typeof address === "string") {
-        reject(new Error("server has no address"));
-        return;
-      }
-      resolve(address);
-    });
-  });
-}
-
-function closeServer(server: http.Server): Promise<void> {
-  return new Promise((resolve, reject) => {
-    server.close((error) => {
-      if (error) reject(error);
-      else resolve();
-    });
-  });
-}
-
-function waitForOpen(ws: WebSocket, timeoutMs = 5_000): Promise<void> {
-  if (ws.readyState === WebSocket.OPEN) return Promise.resolve();
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => reject(new Error("WebSocket open timed out")), timeoutMs);
-    ws.once("open", () => {
-      clearTimeout(timer);
-      resolve();
-    });
-  });
-}
-
-function waitForClose(ws: WebSocket, timeoutMs = 5_000): Promise<number> {
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => reject(new Error("WebSocket close timed out")), timeoutMs);
-    ws.once("close", (code) => {
-      clearTimeout(timer);
-      resolve(code);
-    });
-  });
-}
-
-function connectClient(port: number): WebSocket {
-  const client = new WebSocket(`ws://127.0.0.1:${port}/ws`);
-  client.on("error", () => {});
-  return client;
-}
-
-function websocketUpgradeRequest(port: number): Buffer {
-  const key = Buffer.from("1234567890abcdef").toString("base64");
-  return Buffer.from(
-    `GET /ws HTTP/1.1\r\n` +
-      `Host: 127.0.0.1:${port}\r\n` +
-      `Upgrade: websocket\r\n` +
-      `Connection: Upgrade\r\n` +
-      `Sec-WebSocket-Key: ${key}\r\n` +
-      `Sec-WebSocket-Version: 13\r\n` +
-      `\r\n`,
-  );
-}
-
-function oversizedMaskedBinaryFrame(payloadLength: number): Buffer {
-  const header = Buffer.alloc(2 + 8 + 4);
-  header[0] = 0x82;
-  header[1] = 0x80 | 127;
-  header.writeBigUInt64BE(BigInt(payloadLength), 2);
-  return Buffer.concat([header, Buffer.alloc(payloadLength)]);
-}
-
-async function waitUntil(predicate: () => boolean, timeoutMs = 5_000): Promise<void> {
-  const started = Date.now();
-  while (Date.now() - started < timeoutMs) {
-    if (predicate()) return;
-    await new Promise((resolve) => setTimeout(resolve, 10));
-  }
-  throw new Error("waitUntil timed out");
-}
 
 function waitForExit(
   child: ChildProcess,
@@ -345,20 +94,16 @@ async function runWebsocketProbe(mode: string): Promise<{
     CONFIG_PATH: configPath,
     BACKEND_URL: "http://127.0.0.1:9",
     FRONTEND_BACKEND_API_KEY: DUMMY_API_KEY,
-    FRONTEND_ROOT,
     WEBSOCKET_PROBE_MODE: mode,
+    TSX_TSCONFIG_PATH: path.join(FRONTEND_ROOT, "tsconfig.vite.json"),
   };
 
   try {
-    child = spawn(
-      process.execPath,
-      ["--import", "tsx", "--input-type=module", "--eval", PROBE_SOURCE],
-      {
-        cwd: FRONTEND_ROOT,
-        env: childEnv,
-        stdio: ["ignore", "pipe", "pipe"],
-      },
-    );
+    child = spawn(process.execPath, ["--import", "tsx", PROBE_PATH], {
+      cwd: FRONTEND_ROOT,
+      env: childEnv,
+      stdio: ["ignore", "pipe", "pipe"],
+    });
     child.stdout?.setEncoding("utf8");
     child.stderr?.setEncoding("utf8");
     child.stdout?.on("data", (chunk: string) => {
@@ -483,6 +228,7 @@ describe("oversized pre-auth websocket child probes", () => {
 describe("accepted browser websocket error handling", () => {
   const servers: Array<{ httpServer: http.Server; wss: WebSocketServer }> = [];
   const clients: WebSocket[] = [];
+  const rawSockets: net.Socket[] = [];
 
   beforeEach(() => {
     resetClientErrorLogThrottleForTests();
@@ -490,6 +236,9 @@ describe("accepted browser websocket error handling", () => {
 
   afterEach(async () => {
     resetClientErrorLogThrottleForTests();
+    for (const socket of rawSockets.splice(0)) {
+      socket.destroy();
+    }
     for (const client of clients.splice(0)) {
       client.terminate();
     }
@@ -664,6 +413,7 @@ describe("accepted browser websocket error handling", () => {
     servers.push(started);
 
     const socket = net.connect(started.port, "127.0.0.1");
+    rawSockets.push(socket);
     socket.on("error", () => {});
     await new Promise<void>((resolve, reject) => {
       socket.once("connect", () => resolve());
